Fix for failing UI tests
